+++ ndonnx/_data_types/__init__.py
@@ -2,7 +2,7 @@
 # SPDX-License-Identifier: BSD-3-Clause
 
 from __future__ import annotations
-from warnings import warn
+from ndonnx._utility import deprecated
 
 from .aliases import (
     bool,
@@ -51,9 +51,8 @@ from .schema import Schema
 from .structtype import StructType
 
 
-# TODO: to be removed
-def promote_nullable(dtype: StructType | CoreType) -> _NullableCore:
-    """Promotes a non-nullable type to its nullable counterpart, if present.
+def into_nullable(dtype: StructType | CoreType) -> _NullableCore:
+    """Return nullable counterpart, if present.
 
     Parameters
     ----------
@@ -70,13 +69,6 @@ def promote_nullable(dtype: StructType | CoreType) -> _NullableCore:
     ValueError
         If the input type is unknown to ``ndonnx``.
     """
-
-    warn(
-        "Function 'ndonnx.promote_nullable' will be deprecated in ndonnx 0.7. "
-        "To create nullable array, use 'ndonnx.additional.make_nullable' instead.",
-        DeprecationWarning,
-    )
-
     if dtype == bool:
         return nbool
     elif dtype == float32:
@@ -105,6 +97,14 @@ def promote_nullable(dtype: StructType | CoreType) -> _NullableCore:
         return dtype
     else:
         raise ValueError(f"Cannot promote {dtype} to nullable")
+
+
+@deprecated(
+    "Function 'ndonnx.promote_nullable' will be deprecated in ndonnx 0.7. "
+    "To create nullable array, use 'ndonnx.additional.make_nullable' instead."
+)
+def promote_nullable(dtype: StructType | CoreType) -> _NullableCore:
+    return into_nullable(dtype)
 
 
 __all__ = [